refactor: PostsManagerPage에서 useQueryGetPosts로 posts 가져오게 변경
