+++ src/entities/post/model/types/post.ts
@@ -23,6 +23,4 @@ export interface PostWithUser extends Post {
 export interface PostsResponse {
   posts: Post[]
   total: number
-  skip: number
-  limit: number
 }
